chore: split exchange.rs, it became too big
