+++ crates/tests/fixtures/dispatcher/verbs-worker.php
@@ -159,6 +159,10 @@ try {
             unset($ex); // head already on the wire: the host can only truncate
             continue;
         }
+        if ($probe === 'bail-with-unit') {
+            @trigger_error('bail with unit out', E_USER_ERROR); // bailout: the unit dies with the cycle
+            continue;
+        }
         if ($probe === 'destruct-explicit') {
             $ex->__destruct(); // explicit call is a no-op on a live unit
             $ex->writeBody('explicit-destruct-ok');
